// src/dorc-web/src/apis/dorc-api/models/DatabaseAccessState.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type {
    DatabaseAccessMembership,
    DatabaseAccessPrincipal,
    DatabaseAccessRole,
} from './';

/**
 * @export
 * @interface DatabaseAccessState
 */
export interface DatabaseAccessState {
    /**
     * @type {number}
     * @memberof DatabaseAccessState
     */
    DatabaseId: number;
    /**
     * @type {number}
     * @memberof DatabaseAccessState
     */
    Revision: number;
    /**
     * @type {string}
     * @memberof DatabaseAccessState
     */
    Provider: string;
    /**
     * @type {Array<DatabaseAccessPrincipal>}
     * @memberof DatabaseAccessState
     */
    Principals: Array<DatabaseAccessPrincipal>;
    /**
     * @type {Array<DatabaseAccessRole>}
     * @memberof DatabaseAccessState
     */
    Roles: Array<DatabaseAccessRole>;
    /**
     * @type {Array<DatabaseAccessMembership>}
     * @memberof DatabaseAccessState
     */
    Memberships: Array<DatabaseAccessMembership>;
}